Enables Linux build support
Adds support for building the project on Linux.

This includes modifying the build script to detect the
OS and use appropriate build commands, making the code
platform-independent via preprocessor directives, and
setting position independent code in CMake.

+++ src/fs/api_fs.cpp
@@ -66,7 +66,7 @@ uint32_t FS_Search_GetCount(void* object) {
     return result;
 }
 
-uint32_t FS_Search_GetPathW(void* object, uint32_t index, wchar_t* buf, size_t buf_len) {
+uint32_t FS_Search_GetPathW(void* object, uint32_t index, wchar_t* buf, uint32_t buf_len) {
     uint32_t result = 0;
     if(!object) {
         return result;
